Added a mute command, but it's not what I wanted
I'm only adding this for later but I created a mute command using Discord.js's feature for muting users but apparently that only mutes them on VC, not in text. I need to add some code for creating a role and making that role mute in every channel so that it can be assigned when a user needs to be muted.